рефакторинг метода с утилитами
