+++ src/mock/offer.js
@@ -1,4 +1,4 @@
-import {getRandomInteger} from '../utils.js';
+import {getRandomInteger} from '../utils/common.js';
 import {OFFER_LIST} from '../const.js';
 
 /**
