+++ src/mock/trip-info.js
@@ -1,4 +1,5 @@
-import {getRandomInteger, getRandomDayAndTime, getRandomDestination} from '../utils.js';
+import {getRandomInteger, getRandomDayAndTime} from '../utils/common.js';
+import {getRandomDestination} from '../utils/trip.js';
 import {MIN_PRICE, MAX_PRICE, MIN_ROUTE_POINT_COUNT, MAX_ROUTE_POINT_COUNT} from '../const.js';
 /**
  * Функция геренации данных c информацией и общей маршруте
